refactor: extract storedChunkToSegmentChunk() shared helper (#233)
Prevents field-dropping bugs when rebuilding segments. Multiple commands
(reingest, materialize-edges) independently mapped stored chunk fields
and each forgot optional Chunk fields like documentId/contentFingerprint.

Now there's a single source of truth in segment-builder.ts. Both call
sites replaced with one-liner calls. When new Chunk fields are added,
only this helper needs updating.

Removes reingest-metadata.test.ts (superseded by segment-builder tests).
